Spelling and grammatical fixes in Readme.md
As requested in Issue #4 

I changed the first 5 points to questions to make them more consistent.

Also, I changed the points in "10 Reasons Why Quaco Might Interest You" to be more about the qualities of Quaco and what it can do for you rather than ideals such as with the the 2nd, 3rd, and 4th points. The ideals are covered under the "4 Rules" topic anyways.
I'm not sure about the 7th point though. Are you saying Wordpress doesn't properly support the development of real time themes? If so it should be "The development of real-time themes (by means of XML requests and PHP / HTML) is not yet properly supported by Wordpress".